Refatorando JwtModule de app.module para seguranca.module
