+++ src/app/app.module.ts
@@ -4,17 +4,11 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
 
-import { JwtModule } from '@auth0/angular-jwt';
-
 import { CoreModule } from './core/core.module';
 import { LancamentosModule } from './lancamentos/lancamentos.module';
 import { PessoasModule } from './pessoas/pessoas.module';
 import { AppRoutingModule } from './app-routing.module';
 import { SegurancaModule } from './seguranca/seguranca.module';
-
-export function tokenGetter() {
-  return localStorage.getItem('token');
-}
 
 @NgModule({
   declarations: [AppComponent],
@@ -27,14 +21,7 @@ export function tokenGetter() {
     LancamentosModule,
     PessoasModule,
     AppRoutingModule,
-    SegurancaModule,
-    JwtModule.forRoot({
-      config: {
-        tokenGetter: tokenGetter,
-        // whitelistedDomains: ['localhost:8080'],
-        blacklistedRoutes: ['oauth/token']
-      }
-    })
+    SegurancaModule
   ],
   providers: [],
   bootstrap: [AppComponent]
