+++ src/app/seguranca/seguranca.module.ts
@@ -2,19 +2,33 @@ import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { JwtModule } from '@auth0/angular-jwt';
 import { InputTextModule } from 'primeng/inputtext';
 import { ButtonModule } from 'primeng/button';
 
 import { SegurancaRoutingModule } from './seguranca-routing.module';
 import { LoginFormComponent } from './login-form/login-form.component';
 
+export function tokenGetter() {
+  return localStorage.getItem('token');
+}
+
 @NgModule({
   declarations: [LoginFormComponent],
   imports: [
     CommonModule,
     FormsModule,
+
+    JwtModule.forRoot({
+      config: {
+        tokenGetter: tokenGetter,
+        // whitelistedDomains: ['localhost:8080'],
+        blacklistedRoutes: ['oauth/token']
+      }
+    }),
     InputTextModule,
     ButtonModule,
+
     SegurancaRoutingModule
   ]
 })
